Refactored LogIn feature into a service
