+++ app/src/main/java/com/vmware/nimbus/api/LogInCallback.java
@@ -1,0 +1,6 @@
+package com.vmware.nimbus.api;
+
+public interface LogInCallback {
+    void onSuccess(boolean result);
+    void onFailure(boolean result);
+}
