+++ app/src/main/java/com/vmware/nimbus/ui/login/LoginActivity.java
@@ -26,6 +26,8 @@ import com.android.volley.VolleyError;
 import com.android.volley.toolbox.StringRequest;
 import com.android.volley.toolbox.Volley;
 import com.google.gson.Gson;
+import com.vmware.nimbus.api.APIService;
+import com.vmware.nimbus.api.LogInCallback;
 import com.vmware.nimbus.api.SingletonRequest;
 import com.vmware.nimbus.data.model.CspResult;
 import com.vmware.nimbus.data.model.LoginModel;
@@ -38,7 +40,6 @@ import java.util.Map;
 public class LoginActivity extends AppCompatActivity {
 
     String LOG_TAG = "LoginActivity";
-    private String cspUrl = "https://console.cloud.vmware.com/csp/gateway/am/api/auth/api-tokens/authorize";
 
     @Override
     public void onCreate(Bundle savedInstanceState) {
@@ -57,49 +58,65 @@ public class LoginActivity extends AppCompatActivity {
             public void onClick(View v) {
                 loadingProgressBar.setVisibility(View.VISIBLE);
 
-                StringRequest jsonObjRequest = new StringRequest(
-
-                        Request.Method.POST,
-                        cspUrl,
-                        new Response.Listener<String>() {
+                APIService.LogIn(getBaseContext(), getResources().getString(R.string.cps_URL),
+                        apiKeyEditText.getText().toString(), new LogInCallback() {
                             @Override
-                            public void onResponse(String response) {
-                                Log.d("volley response", response);
-                                Gson gson = new Gson();
-                                CspResult cspResult = gson.fromJson(response, CspResult.class);
-                                LoginModel.getInstance(getBaseContext()).setAuthenticated(true);
-                                LoginModel.getInstance(getBaseContext()).setApi_token(apiKeyEditText.getText().toString());
-                                LoginModel.getInstance(getBaseContext()).setBearer_token(cspResult.getAccess_token());
+                            public void onSuccess(boolean result) {
                                 startActivity(mainIntent);
                                 //Complete and destroy login activity once successful
                                 finish();
                             }
-                        },
-                        new Response.ErrorListener() {
+
                             @Override
-                            public void onErrorResponse(VolleyError error) {
-                                Log.d("volley", "Error: " + error.getMessage());
-                                error.printStackTrace();
+                            public void onFailure(boolean result) {
                                 loadingProgressBar.setVisibility(View.INVISIBLE);
                                 toastMsg("Login Failed");
                             }
-                        }) {
+                        });
 
-                    @Override
-                    public String getBodyContentType() {
-                        return "application/x-www-form-urlencoded; charset=UTF-8";
-                    }
-
-                    @Override
-                    protected Map<String, String> getParams() throws AuthFailureError {
-                        Map<String, String> params = new HashMap<String, String>();
-                        params.put("refresh_token", apiKeyEditText.getText().toString());
-                        return params;
-                    }
-                };
-
-                SingletonRequest.getInstance(getBaseContext()).addToRequestQueue(jsonObjRequest);
-                Log.d(LOG_TAG, "Added request to queue.");
+//                StringRequest jsonObjRequest = new StringRequest(
+//
+//                        Request.Method.POST,
+//                        getResources().getString(R.string.cps_URL),
+//                        new Response.Listener<String>() {
+//                            @Override
+//                            public void onResponse(String response) {
+//                                Log.d("volley response", response);
+//                                Gson gson = new Gson();
+//                                CspResult cspResult = gson.fromJson(response, CspResult.class);
+//                                LoginModel.getInstance(getBaseContext()).setAuthenticated(true);
+//                                LoginModel.getInstance(getBaseContext()).setApi_token(apiKeyEditText.getText().toString());
+//                                LoginModel.getInstance(getBaseContext()).setBearer_token(cspResult.getAccess_token());
+//                                startActivity(mainIntent);
+//                                //Complete and destroy login activity once successful
+//                                finish();
+//                            }
+//                        },
+//                        new Response.ErrorListener() {
+//                            @Override
+//                            public void onErrorResponse(VolleyError error) {
+//                                Log.d("volley", "Error: " + error.getMessage());
+//                                error.printStackTrace();
+//                                loadingProgressBar.setVisibility(View.INVISIBLE);
+//                                toastMsg("Login Failed");
+//                            }
+//                        }) {
+//
+//                    @Override
+//                    public String getBodyContentType() {
+//                        return "application/x-www-form-urlencoded; charset=UTF-8";
+//                    }
+//
+//                    @Override
+//                    protected Map<String, String> getParams() throws AuthFailureError {
+//                        Map<String, String> params = new HashMap<String, String>();
+//                        params.put("refresh_token", apiKeyEditText.getText().toString());
+//                        return params;
+//                    }
+//                };
+//
+//                SingletonRequest.getInstance(getBaseContext()).addToRequestQueue(jsonObjRequest);
+//                Log.d(LOG_TAG, "Added request to queue.");
             }
         });
     }
