+++ app/src/main/res/values/strings.xml
@@ -28,4 +28,6 @@
 
     <string name="world">world</string>
     <string name="contact_dev">Contact the Dev Team</string>
+
+    <string name="cps_URL">https://console.cloud.vmware.com/csp/gateway/am/api/auth/api-tokens/authorize</string>
 </resources>
